Update list_nds_compar.R
Code refactoring. A new function "nds_compar" is included,  generating the comparison for a particular pair of graphs given in the parameter grphs. The parameter grphs is suposed to be a list with two graphs. (For the moment, no check is included. If more than two graphs, the rest will be ignored. If less than two, it is going to be a non-handled error.)

We could consider this function as internal. But I think that it is useful for general users. If you only need to compare 2 graphs, there is no need to generate all pairs.

Additional comment: I think that the assigment of the colors and sizes for the node labels should be done in the plot function, instead of here. I could make this change in the code if you agree. But I would need help to localize all places where the examples and documentation should be changed.