Update list_compar.Rd
Comment: The version that I included is also valid to compare edges of directed graphs. I have just removed the mention of the graphs to be undirected. We could also mention explicitly that the comparison is valid for directed and undirected. But I don't know if this is of any expected usage. Perhaps in other application areas.