Update list_compar.R
Changed parameter var to nds.var, which always refers to nodes. The edge types are always at "type".

Inclusion of directed nature of edges of type ">" or "+". Now an edge A-->--B is not equal to B-->--A, and A--+--B is not equal to B--+--A. 
In contrast, A--=--B is equal to B--=--A.