Update list_compar.R
nds.var is attached in the output of list_compar for each pairwise comparison. This allows the output to keep information about which nodes "type" has been used for the comparison. Thus, the plot_compar function does not need to be provided nds.var as an independent parameter, making the full pipeline more robust.