Update plot_compar.R
Added the number of nodes/edges coincidences in the figure caption.

Removed the plot_compar parameter nds.var.
nds.var is now attached in the output of list_compar for each pairwise comparison. This allows the output to keep information about which nodes "type" has been used for the comparison. Thus, the plot_compar function does not need to be provided nds.var as an independent parameter, making the full pipeline more robust.