Update and rename label_shadow.R to labels_shadow.R
Name of the file in agreement with name of the function.

Plotting the background with a single vectorial graphics::text, avoiding the for loop. This makes it much faster, at least, when used in quartz device. Surprisingly, there is no difference when plotting directly to a file in any format (checked png and pdf).