Update contemp_nds.R
Bug correction. The previous version worked correctly for nodes but not for edges. It assumed that all edges with "=" where contemporaneous to any node.
It worked well in the example graph because it had only one node (1) from a different time than the rest (2, 3, 4). But it only worked if one asked for the contemporanous of the large group (select.nd = 2, for instance). However, for select.nd = 1, the returned edges were (erroneously) the same as for select.nd = 2.

The new version corrects this misbehaviour.