Update plot_compar.Rd
Integration of plot_dec_grph with plot_compar. Now they both use the same internal functions. This represented a big refactoring.
Accordingly, text sizes had to be updated to a different default. (The two plotting functions behaved quite differently before).

img is now directly an input. The previous version read it internally and silently from file "imgs.tsv". This was not very logical and was a big documentation problem: It forced the file including the images dataset to be named "imgs.tsv", to which all users should comply. And it required it to be included in tsv format. A normal user will use either csv or tsv, but not necessarily both.

Added also in both plotting functions the option to plot directly in the active device. This is obtained by img.format = NULL. The default is still img.format = "png", but I would propose to make NULL the default.

Changed parameter name from "doss" to "dir". I think that it is a more standard name. Please, tell me if you don't agree. This change is in all functions and documentations.

Changed parameter name from "graph2" to "dec2comp". It seems more descriptive to me.