Update read_eds.Rd
Changed parameter name from dev to format. It is actually the file format. I don't think that the term "device" is used for that.
Besides, I think that the user should not include the dot for the extension. I have removed it.
Updated the help pages in agreement with this. 
Vignettes doesn't need any update since the paramenter is never used.