Update index.Rmd
Change image_annotate for standard graphics:text. This avoids the compain of github check in linux.
I am not sure if should change this just because of this. Because the error only occurs in github. Not in any other test. Thus I don't think it is necessary for the CRAN.