Update plot_compar.R
Handling also the closing of plot devices in case of premature exit.
This has also corrected a bug which made a new plot window unecessarily open when no one existed even if the output was an image file.
For the proper handling of all pairs I had to introduce a new internal function: plot_pair. We can think if this one could be a public function in the future.